Move FluidScript inside Fluid folder
