+++ src/fluid/widget.js
@@ -3,8 +3,8 @@ import bonuses from './bonuses.js';
 export default {
 	userId: '10001',
 	sessionId: '10000-a-valid-session',
-	locale: 'en-GB',
-	countryCode: 'IRL',
+	locale: 'es-CO',
+	countryCode: 'COL',
 	currencyCode: 'COP',
 	userData: {
 		age: 33,
